+++ src/algo/dfs_pred.rs
@@ -194,7 +194,6 @@ impl<'a, D> DfsPred<'a, D> {
     ///
     /// # Panics
     ///
-    /// * Panics if the `self.next` panics.
     /// * Panics if a source vertex isn't in the digraph.
     /// * Panics if a successor vertex isn't in the digraph.
     ///