+++ src/repr/adjacency_map/fixture.rs
@@ -36,9 +36,9 @@
 //!
 //! ![Kattis, escapewallmaria, sample 3](https://raw.githubusercontent.com/bsdrks/graaf-images/main/out/kattis_escapewallmaria_3.svg)
 
-use {
-    crate::AdjacencyMap,
-    std::collections::BTreeSet,
+use crate::{
+    repr::adjacency_list::fixture,
+    AdjacencyMap,
 };
 
 /// Jørgen Bang-Jensen and Gregory Z. Gutin. 2009. Digraphs: Theory,
@@ -48,16 +48,7 @@ use {
 /// ![Bang-Jensen, 196](https://raw.githubusercontent.com/bsdrks/graaf-images/main/out/bang_jensen_196.svg)
 #[must_use]
 pub fn bang_jensen_196() -> AdjacencyMap {
-    AdjacencyMap::from([
-        BTreeSet::from([1, 4, 7]),
-        BTreeSet::from([0, 2, 7]),
-        BTreeSet::from([3]),
-        BTreeSet::from([2, 4]),
-        BTreeSet::from([2]),
-        BTreeSet::from([6]),
-        BTreeSet::from([7]),
-        BTreeSet::from([5]),
-    ])
+    AdjacencyMap::from(fixture::bang_jensen_196())
 }
 
 /// Jørgen Bang-Jensen and Gregory Z. Gutin. 2009. Digraphs: Theory,
@@ -67,14 +58,7 @@ pub fn bang_jensen_196() -> AdjacencyMap {
 /// ![Bang-Jensen, 34](https://raw.githubusercontent.com/bsdrks/graaf-images/main/out/bang_jensen_34.svg)
 #[must_use]
 pub fn bang_jensen_34() -> AdjacencyMap {
-    AdjacencyMap::from([
-        BTreeSet::from([4]),
-        BTreeSet::from([0]),
-        BTreeSet::from([1, 3, 5]),
-        BTreeSet::new(),
-        BTreeSet::new(),
-        BTreeSet::from([4]),
-    ])
+    AdjacencyMap::from(fixture::bang_jensen_34())
 }
 
 /// Jørgen Bang-Jensen and Gregory Z. Gutin. 2009. Digraphs: Theory,
@@ -84,15 +68,7 @@ pub fn bang_jensen_34() -> AdjacencyMap {
 /// ![Bang-Jensen, 94](https://raw.githubusercontent.com/bsdrks/graaf-images/main/out/bang_jensen_94.svg)
 #[must_use]
 pub fn bang_jensen_94() -> AdjacencyMap {
-    AdjacencyMap::from([
-        BTreeSet::from([1, 2]),
-        BTreeSet::from([3]),
-        BTreeSet::from([1, 3, 4, 5]),
-        BTreeSet::from([5]),
-        BTreeSet::from([6]),
-        BTreeSet::new(),
-        BTreeSet::new(),
-    ])
+    AdjacencyMap::from(fixture::bang_jensen_94())
 }
 
 /// Jeroen Bransen. 2015. Build Dependencies. Kattis.
@@ -101,14 +77,7 @@ pub fn bang_jensen_94() -> AdjacencyMap {
 /// ![Kattis, builddeps](https://raw.githubusercontent.com/bsdrks/graaf-images/main/out/kattis_builddeps.svg)
 #[must_use]
 pub fn kattis_builddeps() -> AdjacencyMap {
-    AdjacencyMap::from([
-        BTreeSet::from([3, 4]),
-        BTreeSet::new(),
-        BTreeSet::from([3, 4, 5]),
-        BTreeSet::from([1]),
-        BTreeSet::from([1]),
-        BTreeSet::from([1]),
-    ])
+    AdjacencyMap::from(fixture::kattis_builddeps())
 }
 
 /// David Sturgill. 2015. Cantina of Babel. (Sample Input 1). Kattis.
@@ -117,20 +86,7 @@ pub fn kattis_builddeps() -> AdjacencyMap {
 /// ![Kattis, cantinaofbabel, sample 1](https://raw.githubusercontent.com/bsdrks/graaf-images/main/out/kattis_cantinaofbabel_1.svg)
 #[must_use]
 pub fn kattis_cantinaofbabel_1() -> AdjacencyMap {
-    AdjacencyMap::from([
-        BTreeSet::from([1]),
-        BTreeSet::from([0, 2, 4]),
-        BTreeSet::from([1]),
-        BTreeSet::from([2, 4, 5, 7, 10, 11]),
-        BTreeSet::from([3]),
-        BTreeSet::from([6]),
-        BTreeSet::from([5, 10]),
-        BTreeSet::from([3]),
-        BTreeSet::from([7, 10]),
-        BTreeSet::from([7, 11]),
-        BTreeSet::from([6]),
-        BTreeSet::from([9]),
-    ])
+    AdjacencyMap::from(fixture::kattis_cantinaofbabel_1())
 }
 
 /// David Sturgill. 2015. Cantina of Babel. (Sample Input 2). Kattis.
@@ -139,20 +95,7 @@ pub fn kattis_cantinaofbabel_1() -> AdjacencyMap {
 /// ![Kattis, cantinaofbabel, sample 2](https://raw.githubusercontent.com/bsdrks/graaf-images/main/out/kattis_cantinaofbabel_2.svg)
 #[must_use]
 pub fn kattis_cantinaofbabel_2() -> AdjacencyMap {
-    AdjacencyMap::from([
-        BTreeSet::from([1]),
-        BTreeSet::from([0, 7]),
-        BTreeSet::from([0, 5, 7]),
-        BTreeSet::from([4]),
-        BTreeSet::from([3]),
-        BTreeSet::from([3, 6]),
-        BTreeSet::from([5]),
-        BTreeSet::from([2]),
-        BTreeSet::from([7, 9, 11]),
-        BTreeSet::from([8]),
-        BTreeSet::from([9, 11]),
-        BTreeSet::from([10]),
-    ])
+    AdjacencyMap::from(fixture::kattis_cantinaofbabel_2())
 }
 
 /// Arash Behpour. 2019. Escape Wall Maria. Kattis. (Sample Input 1)
@@ -161,24 +104,7 @@ pub fn kattis_cantinaofbabel_2() -> AdjacencyMap {
 /// ![Kattis, escapewallmaria, sample 1](https://raw.githubusercontent.com/bsdrks/graaf-images/main/out/kattis_escapewallmaria_1.svg)
 #[must_use]
 pub fn kattis_escapewallmaria_1() -> AdjacencyMap {
-    AdjacencyMap::from([
-        BTreeSet::new(),
-        BTreeSet::new(),
-        BTreeSet::new(),
-        BTreeSet::new(),
-        BTreeSet::new(),
-        BTreeSet::from([6, 9]),
-        BTreeSet::from([5]),
-        BTreeSet::new(),
-        BTreeSet::new(),
-        BTreeSet::from([5, 13]),
-        BTreeSet::new(),
-        BTreeSet::new(),
-        BTreeSet::new(),
-        BTreeSet::from([9, 12]),
-        BTreeSet::new(),
-        BTreeSet::new(),
-    ])
+    AdjacencyMap::from(fixture::kattis_escapewallmaria_1())
 }
 
 /// Arash Behpour. 2019. Escape Wall Maria. Kattis. (Sample Input 2)
@@ -187,24 +113,7 @@ pub fn kattis_escapewallmaria_1() -> AdjacencyMap {
 /// ![Kattis, escapewallmaria, sample 2](https://raw.githubusercontent.com/bsdrks/graaf-images/main/out/kattis_escapewallmaria_2.svg)
 #[must_use]
 pub fn kattis_escapewallmaria_2() -> AdjacencyMap {
-    AdjacencyMap::from([
-        BTreeSet::new(),
-        BTreeSet::new(),
-        BTreeSet::new(),
-        BTreeSet::new(),
-        BTreeSet::new(),
-        BTreeSet::from([6, 9]),
-        BTreeSet::from([5]),
-        BTreeSet::new(),
-        BTreeSet::new(),
-        BTreeSet::from([5]),
-        BTreeSet::new(),
-        BTreeSet::new(),
-        BTreeSet::from([13]),
-        BTreeSet::from([9, 12]),
-        BTreeSet::new(),
-        BTreeSet::new(),
-    ])
+    AdjacencyMap::from(fixture::kattis_escapewallmaria_2())
 }
 
 /// Arash Behpour. 2019. Escape Wall Maria. Kattis. (Sample Input 3)
@@ -213,22 +122,5 @@ pub fn kattis_escapewallmaria_2() -> AdjacencyMap {
 /// ![Kattis, escapewallmaria, sample 3](https://raw.githubusercontent.com/bsdrks/graaf-images/main/out/kattis_escapewallmaria_3.svg)
 #[must_use]
 pub fn kattis_escapewallmaria_3() -> AdjacencyMap {
-    AdjacencyMap::from([
-        BTreeSet::new(),
-        BTreeSet::from([2, 5]),
-        BTreeSet::from([1, 6]),
-        BTreeSet::new(),
-        BTreeSet::new(),
-        BTreeSet::from([1, 6, 9]),
-        BTreeSet::from([2, 5]),
-        BTreeSet::new(),
-        BTreeSet::new(),
-        BTreeSet::from([5, 13]),
-        BTreeSet::new(),
-        BTreeSet::new(),
-        BTreeSet::from([13]),
-        BTreeSet::from([9, 12]),
-        BTreeSet::new(),
-        BTreeSet::new(),
-    ])
+    AdjacencyMap::from(fixture::kattis_escapewallmaria_3())
 }